requirements.txt
openai>=1.30.0
pyyaml>=6.0
python-dotenv>=1.0
gitpython>=3.1
jsonschema>=4.0
